Use describe/it, move test into flow version folder
