+++ definitions/npm/lolex_v2.x.x/flow_v0.64.x-/test_lolex_v2.x.x.js
@@ -1,0 +1,31 @@
+// @flow
+import { describe, it } from 'flow-typed-test';
+import lolex from 'lolex';
+
+describe('lolex', () => {
+  it('install', () => {
+    lolex.install({
+      target: global,
+      now: 150,
+      loopLimit: 250,
+    });
+
+    // $ExpectError
+    lolex.install(true);
+  });
+
+  it('clock', () => {
+    const clock = lolex.install();
+
+    clock.next();
+    clock.tick(150);
+    clock.runAll();
+    const timeoutId = clock.setTimeout(() => {});
+    const intervalId = clock.setInterval(() => {});
+    clock.clearTimeout(timeoutId);
+    // $ExpectError
+    clock.clearTimeout(intervalId);
+    clock.clearInterval(intervalId);
+    clock.runToFrame();
+  });
+});
