--- definitions/npm/lolex_v2.x.x/test_lolex_v2.x.x.js
@@ -1,24 +1,0 @@
-// @flow
-import lolex from 'lolex';
-
-lolex.install({
-  target: global,
-  now: 150,
-  loopLimit: 250,
-});
-
-// $ExpectError
-lolex.install(true);
-
-const clock = lolex.install();
-
-clock.next();
-clock.tick(150);
-clock.runAll();
-const timeoutId = clock.setTimeout(() => {});
-const intervalId = clock.setInterval(() => {});
-clock.clearTimeout(timeoutId);
-// $ExpectError
-clock.clearTimeout(intervalId);
-clock.clearInterval(intervalId);
-clock.runToFrame();
